+++ src/communication/dto/SettingsDTO.ts
@@ -21,6 +21,8 @@ export interface TerminalSettingsDTO {
   cursorBlink?: boolean;
   cursorStyle?: 'block' | 'underline' | 'bar';
   cursorWidth?: number;
+  drawBoldTextInBrightColors?: boolean;
+  minimumContrastRatio?: number;
   scrollback?: number;
   altClickMovesCursor?: boolean;
   highlightActiveBorder?: boolean;